All useful changes of "float" into "double" realized. A verification should be performed to resolve some errors caused by that.

+++ core/math/src/main/java/org/arakhne/afc/math/continous/object3d/Tuple3f.java
@@ -689,9 +689,9 @@ public class Tuple3f<T extends Tuple3D<? super T>> implements Tuple3D<T> {
 	@Override
 	public int hashCode() {
 		int bits = 1;
-		bits = 31 * bits + Double.floatToIntBits(this.x);
-		bits = 31 * bits + Double.floatToIntBits(this.y);
-		bits = 31 * bits + Double.floatToIntBits(this.z);
+		bits = 31 * bits + Float.floatToIntBits(this.x);
+		bits = 31 * bits + Float.floatToIntBits(this.y);
+		bits = 31 * bits + Float.floatToIntBits(this.z);
 		return bits ^ (bits >> 32);
 	}
